feat: complete backend authentication and validation

+++ backend/controllers/cards.js
@@ -23,8 +23,17 @@ module.exports.createCard = (req, res, next) => {
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  Card.findByIdAndDelete(req.params.cardId)
+  Card.findById(req.params.cardId)
     .orFail()
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id) {
+        const error = new Error('Você não tem permissão para excluir este card');
+        error.statusCode = 403;
+        throw error;
+      }
+
+      return Card.findByIdAndDelete(req.params.cardId);
+    })
     .then(() => {
       res.status(200).send({
         message: 'Card excluído com sucesso',
